requirements.txt
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
supabase==2.9.0
python-dotenv==1.0.1
pyyaml==6.0.2
structlog==24.4.0
pytest==8.3.3
pytest-cov==5.0.0
ruff==0.7.1
mypy==1.13.0
